refactor(listener): address second review round
- Rename gnrapplisten.py to gnrlisten.py (command: gnr app listen)
- Remove legacy entry point from pyproject.toml scripts
- Move all runtime imports to module level in tests

Ref #666
